Sebastian Commit
- Se realiza cambio de variables primitivas Float a Double de todas las variables declaradas en todas las clases de los tres package.
- se agrega el metodo senX en la clase TaylorClass;
- se escriben instrucciones de prueba del metodo senX en la MainClass

+++ SerieTaylor/src/Mathpackage/MathClass.java
@@ -5,8 +5,8 @@ public class MathClass {
    
     //Metodo potencia que recibe dos parametros x(valor numero)
     //i(numero de veces que se realizara el producto x 
-    public float mPotencia(float numero, int repeticiones){
-        float potencia = 1 ;
+    public double mPotencia(double numero, int repeticiones){
+        double potencia = 1 ;
         for(int k = 1; k <= repeticiones; k++){
             potencia = potencia * numero;
         } 
@@ -14,8 +14,8 @@ public class MathClass {
     }
 
     //Metodo de Factorial
-    public float mFactorial(float n) {
-        float resultado = 1;
+    public double mFactorial(double n) {
+        double resultado = 1;
         while (n != 0) {
             resultado = resultado * n;
             n--;
@@ -23,7 +23,7 @@ public class MathClass {
         return resultado;
     }
 
-    public float fFactorial(int numero) {
+    public double fFactorial(int numero) {
         return (numero == 0 || numero == 1) ? 1 : numero * fFactorial(numero - 1);
     }
 }
